refactor(game): extract scene/object/system construction to game/scene-setup.ts (R2.4)
Move the one-shot construction out of snowglider.ts into `src/game/scene-setup.ts`
as `setupScene()`: scene/renderer/camera manager, the game-over overlay DOM,
lighting, terrain, the avalanche system, trees (+ `addTreesWithPositions`), the
snowman, snow particles, and the course/effects subsystems. The `GameState` type,
the `AVALANCHE_TRIGGER_DISTANCE` constant, and the eager `window.terrainMesh` /
`treePositions` / `rockPositions` / `isTestMode` data globals move with it.

The coordinator now calls `setupScene()` once and destructures the handles the run
loop, lifecycle, and `publishGameGlobals()` proxies need (`scene`, `renderer`,
`camera`, `cameraManager`, overlay nodes, `terrain`, `rockPositions`,
`treePositions`, `snowman`, `snowSplash`, `state`). The player physics state
(`Physics.createPlayerState`) stays in the coordinator per the plan, so
`publishGameGlobals` still reads `player.*` / `pos` / `velocity` locally.

Mechanical, behavior-preserving: construction order and side effects are unchanged;
`publishGameGlobals()` and every published name stay byte-identical. Removed four
now-unused coordinator imports (Camera, AvalancheSystem, RockPosition, TreePosition).

Verification: lint (0 errors) / typecheck / test (contract guard 44/44) / build
green; puppeteer 87/87 across 7/7 suites (camera/regression/tree/avalanche drive the
moved scene globals).
